@@ -4,49 +4,71 @@
   "version": "1.0.0",
   "description": "frontend for the sextant cluster creation service",
   "scripts": {
-    "develop": "webpack-dev-server --mode development --config config/webpack.base.config.js --hot --history-api-fallback --host 0.0.0.0 --disable-host-check",
-    "build": "webpack --mode production --config config/webpack.prod.config.js --env.NODE_ENV=production --progress"
+    "develop": "webpack-dev-server --config config/webpack.dev.js",
+    "build": "webpack --config config/webpack.prod.js",
+    "test": "babel-tape-runner {src,test}/**/*.test.js | tap-spec",
+    "analyze": "npm run analyze:build && npm run analyze:serve",
+    "analyze:build": "webpack --config config/webpack.analyze.js",
+    "analyze:serve": "node analyzeServer.js"
   },
   "dependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.0.0",
-    "@babel/plugin-proposal-decorators": "7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "7.0.0",
-    "@babel/plugin-proposal-function-sent": "7.0.0",
-    "@babel/plugin-proposal-json-strings": "7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "7.0.0",
-    "@babel/plugin-syntax-import-meta": "7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@material-ui/core": "^3.9.1",
-    "@material-ui/icons": "^3.0.2",
-    "@material-ui/lab": "^3.0.0-alpha.30",
-    "axios": "^0.18.0",
-    "babel-loader": "^8.0.2",
-    "copy-webpack-plugin": "^4.5.1",
-    "history": "^4.7.2",
-    "html-webpack-plugin": "^3.2.0",
-    "prop-types": "^15.6.1",
-    "query-string": "^6.1.0",
-    "randomstring": "^1.1.5",
-    "react": "^16.3.0",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.3.0",
-    "react-hot-loader": "^4.1.1",
-    "react-redux": "^4.0.0 || ^5.0.0",
-    "redux": "^3.0.0",
-    "redux-box": "^1.6.1",
-    "redux-first-router": "^0.0.16-next",
-    "redux-first-router-link": "^1.4.2",
-    "redux-form": "^7.3.0",
-    "redux-saga": "^0.16.0",
-    "timeago.js": "^4.0.0-beta.1",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "webpack": "^4.12.1",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-server": "^3.1.4",
-    "webpack-merge": "^4.1.3"
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "7.3.4",
+    "@babel/plugin-proposal-class-properties": "7.3.4",
+    "@babel/plugin-proposal-decorators": "7.3.0",
+    "@babel/plugin-proposal-export-namespace-from": "7.2.0",
+    "@babel/plugin-proposal-function-sent": "7.2.0",
+    "@babel/plugin-proposal-json-strings": "7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "7.2.0",
+    "@babel/plugin-syntax-import-meta": "7.2.0",
+    "@babel/plugin-transform-react-jsx-source": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.4.0",
+    "@babel/preset-env": "7.3.4",
+    "@babel/preset-react": "7.0.0",
+    "@material-ui/core": "3.9.2",
+    "@material-ui/icons": "3.0.2",
+    "autosuggest-highlight": "^3.1.1",
+    "axios": "0.18.0",
+    "babel-loader": "8.0.5",
+    "babel-tape-runner": "^3.0.0",
+    "classnames": "^2.2.6",
+    "clean-webpack-plugin": "^2.0.1",
+    "copy-webpack-plugin": "5.0.0",
+    "dotty": "^0.1.0",
+    "ecstatic": "^3.3.1",
+    "error-overlay-webpack-plugin": "^0.2.0",
+    "formik": "^1.5.1",
+    "html-webpack-plugin": "3.2.0",
+    "normalizr": "^3.3.0",
+    "pretty-bytes": "^5.1.0",
+    "prop-types": "15.7.2",
+    "react": "16.8.4",
+    "react-autosuggest": "^9.4.3",
+    "react-dnd": "^7.4.5",
+    "react-dnd-html5-backend": "^7.4.4",
+    "react-dom": "16.8.4",
+    "react-dropzone": "^10.1.0",
+    "react-redux": "6.0.1",
+    "react-router5": "^7.1.0",
+    "redux": "4.0.1",
+    "redux-router5": "^7.0.2",
+    "redux-saga": "^1.0.2",
+    "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
+    "router5": "^7.0.2",
+    "router5-plugin-browser": "^7.0.2",
+    "router5-transition-path": "^7.0.1",
+    "tap-spec": "^5.0.0",
+    "tape": "^4.10.1",
+    "terser-webpack-plugin": "^1.2.3",
+    "webpack": "4.29.6",
+    "webpack-bundle-analyzer": "^3.1.0",
+    "webpack-cli": "3.2.3",
+    "webpack-dev-server": "3.2.1",
+    "webpack-merge": "4.2.1",
+    "webpack-visualizer-plugin": "^0.1.11",
+    "yup": "^0.27.0"
   }
 }
